requirements.txt
chromadb
openai
python-dotenv
langchain_community
langchain_text_splitters
pypdf
Pillow
requests